* fixed all GameDefinition related stuff (admin doesn't work right now)

+++ web_app/app/models/score.rb
@@ -1,9 +1,9 @@
 class Score < ActiveRecord::Base
-  has_many :fragments, :class_name => "Score::Fragment", :foreign_key => "score_id", :dependent => :destroy
+  
+  validates_inclusion_of :game_definition, :in => %w{HaseUndIgel}
+  validates_inclusion_of :score_type, :in => %w{round_score match_score}
 
-  def score_type
-    :round_score
-  end
+  has_many :fragments, :class_name => "Score::Fragment", :foreign_key => "score_id", :dependent => :destroy
 
   def definition
     GameDefinition.all.first.send(score_type)
@@ -11,13 +11,14 @@ class Score < ActiveRecord::Base
 
   def set!(values)
     raise "values must be an Array" unless values.is_a? Array
+    raise "values length was #{values.size}, expected: #{definition.size}" unless values.size == definition.size
 
     Score.transaction do
       fragments.destroy_all
       save! if new_record?
 
-      values.each do |value|
-        self.fragments.create!(:score => self, :value => values)
+      definition.values.each_with_index do |fragment,i|
+        self.fragments.create!(:score => self, :fragment => fragment.name.to_s, :value => values[i])
       end
       
       save!
@@ -41,11 +42,15 @@ class Score < ActiveRecord::Base
 
     validates_presence_of :score
     validates_presence_of :value
+    
+    validates_presence_of :fragment
+    validates_format_of :fragment, :with => /\A[a-z0-9_]*\Z/
 
     belongs_to :score
 
     def value_with_precision
-      sprintf("%.#{score.definitiondefinition.precision}f", value)
+      precision = score.definition[fragment.to_sym].precision
+      sprintf("%.#{precision}f", value)
     end
   end
 end
